requirements.txt
Django>=1.9.0,<1.8.99

# Contrib ---------------------------------------------------------------------
django-autoslug==1.9.3
Jinja2==2.8
requests==2.9.1
django-memoize==1.2.0

# Assets ----------------------------------------------------------------------
pillow
whitenoise==2.0.6
django-pipeline
Wand


# Production ------------------------------------------------------------------
psycopg2
raven==5.11.1
gunicorn
python-memcached


# Testing ---------------------------------------------------------------------
coverage